+++ sevivi/config/config_types/__init__.py
@@ -1,0 +1,5 @@
+from .plotting_method import PlottingMethod
+from .stacking_direction import StackingDirection
+from .video_config import VideoConfig
+from .config import Config
+from .render_config import RenderConfig
