+++ sevivi/config/config_types/video_config.py
@@ -1,6 +1,6 @@
 import os
 from dataclasses import dataclass
-from typing import Union, List
+from typing import List
 
 
 @dataclass
