+++ sevivi/image_provider/__init__.py
@@ -1,1 +1,7 @@
-from sevivi.image_provider.video_provider import AzureProvider, PlainVideoImageProvider
+from .video_provider import (
+    AzureProvider,
+    PlainVideoImageProvider,
+    ImuCameraImageProvider,
+    VideoImageProvider,
+)
+from .graph_provider import GraphImageProvider
