+++ sevivi/image_provider/graph_provider/graph_provider.py
@@ -1,39 +1,17 @@
-from abc import ABC
-from collections import defaultdict
-from typing import Tuple, List, Optional, Dict, Union
+from typing import List, Optional
 
 import pandas as pd
 from matplotlib.axis import Axis
 
-from sevivi.config import RenderConfig
-from sevivi.config.config_types.sensor_config import (
+from sevivi.config import (
+    RenderConfig,
+    find_matching_columns,
     SensorConfig,
     ManuallySynchronizedSensorConfig,
     JointSynchronizedSensorConfig,
     ImuSynchronizedSensorConfig,
+    get_graph_groups,
 )
-
-
-def get_graph_groups(
-    df: pd.DataFrame, patterns: Optional[List[str]]
-) -> Dict[str, List[str]]:
-    if patterns is None:
-        return {col: [col] for col in df.columns}
-    else:
-        return {pattern: find_matching_columns(df, pattern) for pattern in patterns}
-
-
-def find_matching_columns(
-    df: pd.DataFrame, pattern: Union[str, List[str]]
-) -> List[str]:
-    if isinstance(pattern, str):
-        return [col for col in df.columns if pattern in col]
-    else:
-        column_set = set(df.columns)
-        desired_column_set = set(pattern)
-        if len(desired_column_set - column_set) > 0:
-            raise ValueError(f"Unknown columns {desired_column_set - column_set}")
-        return pattern
 
 
 class GraphImageProvider:
