+++ sevivi/image_provider/video_provider/__init__.py
@@ -1,2 +1,4 @@
 from .azure_image_provider import AzureProvider
 from .plain_video_image_provider import PlainVideoImageProvider
+from .imu_camera_image_provider import ImuCameraImageProvider
+from .video_provider import VideoImageProvider
